main converted to fragment
