+++ frontend/app/src/main/AndroidManifest.xml
@@ -17,9 +17,6 @@
         tools:remove="android:appComponentFactory"
         tools:targetApi="31">
         <activity
-            android:name=".BaseActivity"
-            android:exported="false" />
-        <activity
             android:name=".LoginActivity"
             android:exported="true">
             <intent-filter>
@@ -34,23 +31,13 @@
             android:label="@string/title_activity_articles"
             android:theme="@style/Theme.MyApplication.NoActionBar" />
         <activity
-            android:name=".SubscribedArticlesActivity"
-            android:exported="false"
-            android:label="@string/title_activity_subscribed"
-            android:theme="@style/Theme.MyApplication.NoActionBar" />
+            android:name=".BaseActivity"
+            android:exported="false" />
         <activity
             android:name=".ForumActivity"
             android:exported="false"
             android:label="@string/title_activity_forum"
             android:theme="@style/Theme.MyApplication.NoActionBar" />
-        <activity
-            android:name=".ForumsListActivity"
-            android:exported="false"
-            android:label="@string/title_activity_forums_list"
-            android:theme="@style/Theme.MyApplication.NoActionBar" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="false" />
         <activity
             android:name=".HistoryActivity"
             android:exported="false"
