+++ frontend/app/src/main/res/layout/activity_subscribed_articles.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SubscribedArticlesActivity">
+    tools:context=".SubscribedArticlesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/view_article"
